+++ apple/Features/TimeSeries/TimeSeriesTimestampAppendExample.m
@@ -117,7 +117,7 @@
          }
 
          NSString *const author = eventMetadata.author;
-         NSDate *const timestamp = [NSDate dateWithTimeIntervalSince1970:eventMetadata.timestamp / 1000];
+         NSDate *const timestamp = [NSDate dateWithTimeIntervalSince1970:(double)eventMetadata.timestamp / 1000.0];
          NSDateFormatter *const dateFormatter = [[NSDateFormatter alloc] init];
          [dateFormatter setDateFormat:@"yyyy-MM-dd HH:mm:ss"];
          NSLog(@"Time series topic with user supplied timestamp appended by [%@] at [%@]", author, [dateFormatter stringFromDate:timestamp]);
